requirements_production.txt
# ArcSpatialDB Production Requirements
# Minimal requirements for production deployment

# Core Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Database ORM
SQLAlchemy==2.0.21

# HTTP Client for API calls
requests==2.31.0

# File pattern matching
glob2==0.7

# Production WSGI Server
gunicorn==21.2.0

# CORS support for backend API
Flask-CORS==4.0.0

# Word document processing (for code collection feature)
python-docx==0.8.11

# Alternative WSGI Server for Windows
waitress==2.1.2 